Refactor GUI and updater
